+++ packages/web3-utils/src/json_rpc.ts
@@ -87,14 +87,34 @@ export const isBatchResponse = <Result = unknown, Error = unknown>(
 ): response is JsonRpcBatchResponse<Result, Error> =>
 	Array.isArray(response) && response.length > 1 && isValidResponse(response);
 
+// internal optional variable to increment and use for the jsonrpc `id`
+let requestIdSeed: number | undefined;
+
+/**
+ * Optionally use to make the jsonrpc `id` start from a specific number.
+ * Without calling this function, the `id` will be filled with a Uuid.
+ * But after this being called with a number, the `id` will be a number staring from the provided `start` variable.
+ * However, if `undefined` was passed to this function, the `id` will be a Uuid again.
+ * @param start - a number to start incrementing from.
+ * 	Or `undefined` to use a new Uuid (this is the default behavior)
+ */
+export const setRequestIdStart = (start: number | undefined) => {
+	requestIdSeed = start;
+};
+
 export const toPayload = <ParamType = unknown[]>(
 	request: JsonRpcOptionalRequest<ParamType>,
-): JsonRpcPayload<ParamType> => ({
-	jsonrpc: request.jsonrpc ?? '2.0',
-	id: request.id ?? uuidV4(),
-	method: request.method,
-	params: request.params ?? undefined,
-});
+): JsonRpcPayload<ParamType> => {
+	if (typeof requestIdSeed !== 'undefined') {
+		requestIdSeed += 1;
+	}
+	return {
+		jsonrpc: request.jsonrpc ?? '2.0',
+		id: request.id ?? requestIdSeed ?? uuidV4(),
+		method: request.method,
+		params: request.params ?? undefined,
+	};
+};
 
 export const toBatchPayload = (requests: JsonRpcOptionalRequest<unknown>[]): JsonRpcBatchRequest =>
 	requests.map(request => toPayload<unknown>(request)) as JsonRpcBatchRequest;